gpu bridge: fix Linux + MinGW emulator build (vulkan_host.cpp)
- Linux/GCC (-Werror=extra): the image-view aspect ternary mixed uint32_t with
  the VkImageAspectFlagBits enum; cast VK_IMAGE_ASPECT_COLOR_BIT to uint32_t so
  both branches are the same type.
- MinGW: <Windows.h> -> <windows.h>. MinGW's header is lowercase and the cross-
  build runs on a case-sensitive filesystem, so the capitalized include was a
  fatal not-found error (it had hidden everything else in the TU).

+++ src/windows-emulator/devices/vulkan_host.cpp
@@ -16,7 +16,7 @@
 #ifndef NOMINMAX
 #define NOMINMAX
 #endif
-#include <Windows.h>
+#include <windows.h>
 #else
 #include <dlfcn.h>
 #endif
@@ -2137,7 +2137,7 @@ namespace sogen
         info.image = img->second.handle;
         info.viewType = VK_IMAGE_VIEW_TYPE_2D;
         info.format = static_cast<VkFormat>(format);
-        info.subresourceRange.aspectMask = (aspect_mask != 0) ? aspect_mask : VK_IMAGE_ASPECT_COLOR_BIT;
+        info.subresourceRange.aspectMask = (aspect_mask != 0) ? aspect_mask : static_cast<uint32_t>(VK_IMAGE_ASPECT_COLOR_BIT);
         info.subresourceRange.levelCount = 1;
         info.subresourceRange.layerCount = 1;
